+++ .github/workflows/codeql-analysis.yml
@@ -68,5 +68,8 @@ jobs:
     #   echo "Run, Build Application using script"
     #   ./location_of_script_within_repo/buildscript.sh
 
+    - name: Run make check
+      run: make check
+
     - name: Perform CodeQL Analysis
       uses: github/codeql-action/analyze@v2
