+++ go.mod
@@ -1,11 +1,19 @@
 module github.com/pouriyajamshidi/tcping
 
-go 1.16
+go 1.18
 
 require (
-	github.com/google/go-github/v45 v45.0.0
-	github.com/gookit/color v1.5.0
-	github.com/stretchr/testify v1.7.1
-	golang.org/x/sys v0.0.0-20220520151302-bc2c85ada10a // indirect
+	github.com/google/go-github/v45 v45.1.0
+	github.com/gookit/color v1.5.1
+	github.com/stretchr/testify v1.7.2
+)
+
+require (
+	github.com/davecgh/go-spew v1.1.0 // indirect
+	github.com/google/go-querystring v1.1.0 // indirect
+	github.com/pmezard/go-difflib v1.0.0 // indirect
+	github.com/xo/terminfo v0.0.0-20210125001918-ca9a967f8778 // indirect
+	golang.org/x/crypto v0.0.0-20220525230936-793ad666bf5e // indirect
+	golang.org/x/sys v0.0.0-20220608164250-635b8c9b7f68 // indirect
 	gopkg.in/yaml.v3 v3.0.1 // indirect
 )
